building the advanced classes
