--- Drop-Shipping-Simulation-Uml-Diagram(Project).simp.user
@@ -1,5 +1,0 @@
-﻿<user-preferences>
-  <open-windows>
-    <window type="Diagram" resource="p1117c8ae24a14a0fb9b1c0f3919b9e7b_726706002" />
-  </open-windows>
-</user-preferences>
